Squared difference between standard deviations (SDSD)

It estimates the SDSD component of the Mean Squared Error (MSE) proposed by Kobayashi & Salam (2000).

Usage

SDSD(data = NULL, obs, pred, tidy = FALSE, na.rm = TRUE)

Arguments

data

(Optional) argument to call an existing data frame containing the data.

obs

Vector with observed values (numeric).

pred

Vector with predicted values (numeric).

tidy

Logical operator (TRUE/FALSE) to decide the type of return. TRUE returns a data.frame, FALSE returns a list; Default : FALSE.

na.rm

Logic argument to remove rows with missing values (NA). Default is na.rm = TRUE.

Value

an object of class numeric within a list (if tidy = FALSE) or within a data frame (if tidy = TRUE).

Details

The SDSD represents the proportional bias component of the prediction error following Kobayashi & Salam (2000). It is in square units of the variable of interest, so it does not have a direct interpretation. The lower the value the less contribution to the MSE. However, it needs to be compared to MSE as its benchmark. For the formula and more details, see online-documentation

References

Kobayashi & Salam (2000). Comparing simulated and measured values using mean squared deviation and its components. Agron. J. 92, 345–352. doi:10.2134/agronj2000.922345x

Examples

# \donttest{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
SDSD(obs = X, pred = Y)
#> $SDSD
#> [1] 0.1967212
#> 
# }