metrica: Prediction Performance Metrics

A compilation of more than 80 functions designed to quantitatively and visually evaluate prediction performance of regression (continuous variables) and classification (categorical variables) of point-forecast models (e.g. APSIM, DSSAT, DNDC, supervised Machine Learning). For regression, it includes functions to generate plots (scatter, tiles, density, & Bland-Altman plot), and to estimate error metrics (e.g. MBE, MAE, RMSE), error decomposition (e.g. lack of accuracy-precision), model efficiency (e.g. NSE, E1, KGE), indices of agreement (e.g. d, RAC), goodness of fit (e.g. r, R2), adjusted correlation coefficients (e.g. CCC, dcorr), symmetric regression coefficients (intercept, slope), and mean absolute scaled error (MASE) for time series predictions. For classification (binomial and multinomial), it offers functions to generate and plot confusion matrices, and to estimate performance metrics such as accuracy, precision, recall, specificity, F-score, Cohen's Kappa, G-mean, and many more. For more details visit the vignettes https://adriancorrendo.github.io/metrica/.

See also

Author

Maintainer: Adrian A. Correndo correndo@ksu.edu (ORCID) [copyright holder]

Authors:

Adrian A. Correndo correndo@ksu.edu (ORCID)

Luiz H. Moro Rosso lhmrosso@ksu.edu (ORCID)

Rai Schwalbert rai.schwalbert@hotmail.com (ORCID)

Carlos Hernandez carlosh92@ksu.edu (ORCID)

Leonardo M. Bastos leonardombastos@gmail.com (ORCID)

Luciana Nieto lnieto@ksu.edu (ORCID)

Dean Holzworth dean.holzworth@csiro.au

Ignacio A. Ciampitti ciampitti@ksu.edu (ORCID)