Prediction Performance Summary

It estimates a group of metrics characterizing the prediction performance for a continuous predicted-observed dataset. By default, it calculates all available metrics.

Usage

metrics_summary(
  data = NULL,
  obs,
  pred,
  orientation = "PO",
  na.rm = TRUE,
  metrics_list = NULL
)

Arguments

data

(Optional) argument to call an existing data frame containing the data.

obs

vector with observed values (numeric).

pred

vector with predicted values (numeric).

orientation

argument of class string specifying the axis orientation to estimate slope(B1) and intercept(B0). "PO" is for predicted vs observed, and "OP" for observed vs predicted. Default is orientation = "PO".

na.rm

Logic argument to remove rows with missing values (NA). Default is na.rm = TRUE.

metrics_list

vector or list of specific selected metrics. Default is = NULL, which will estimate all metrics available.

Value

an object of class data.frame containing all (or selected) metrics.

Details

The user can choose to calculate a single metric, or to calculate all metrics at once. This function creates a data.frame with all (or selected) metrics in the metrica-package. If looking for specific metrics, the user can pass a list of desired metrics using the argument “metrics_list” (e.g. metrics_list = c("R2","MAE", "RMSE", "RSR", "NSE", "KGE")). For the entire list of available metrics with formula, see online documentation

Examples

# \donttest{
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 10)
df <- data.frame(obs = X, pred = Y)
metrics_summary(df, obs = X, pred = Y)
#>    Metric        Score
#> 1      B0   0.23150260
#> 2      B1   0.95882067
#> 3       r   0.11078034
#> 4      R2   0.01227228
#> 5      Xa   0.99888019
#> 6     CCC   0.11065629
#> 7     MAE  10.59090806
#> 8    RMAE  35.69142690
#> 9    MAPE 505.61645231
#> 10  SMAPE 131.88269760
#> 11    RAE   1.31926662
#> 12    RSE   1.70735417
#> 13    MBE  -0.21928324
#> 14    PBE -73.89858871
#> 15    PAB   0.02659426
#> 16    PPB   0.09931960
#> 17    MSE 180.81018530
#> 18   RMSE  13.44656779
#> 19  RRMSE  45.31501816
#> 20    RSR   0.12697323
#> 21 iqRMSE   1.16557822
#> 22    MLA   0.22766510
#> 23    MLP 180.58252020
#> 24     SB   0.04808514
#> 25   SDSD   0.17957996
#> 26    LCS 180.58252020
#> 27    PLA   0.12591387
#> 28    PLP  99.87408613
#> 29     Ue  99.87408613
#> 30     Uc   0.09931960
#> 31     Ub   0.02659426
#> 32    NSE  -0.70735417
#> 33     E1  -0.31926662
#> 34   Erel  -3.22490639
#> 35    KGE  -0.24019461
#> 36      d   0.45890828
#> 37     d1   0.33591874
#> 38    d1r   0.33591874
#> 39    RAC   0.55527555
#> 40     AC  -1.61289162
#> 41 lambda   0.11065629
#> 42  dcorr   0.17286650
#> 43    MIC   0.28186524
# }